refactor: remove sse transport
