+++ src/app/cases/__tests__/dragOverlayPosition.test.tsx
@@ -13,14 +13,6 @@ vi.mock("next/navigation", () => ({
 }));
 
 vi.stubGlobal(
-  "EventSource",
-  class {
-    onmessage: ((this: unknown, ev: MessageEvent) => void) | null = null;
-    close() {}
-  },
-);
-
-vi.stubGlobal(
   "fetch",
   vi.fn(async () => ({ ok: true, json: async () => baseCase })),
 );
