+++ src/app/system-status/SystemStatusClient.tsx
@@ -1,6 +1,5 @@
 "use client";
-import { apiEventSource } from "@/apiClient";
-import { subscribe as wsSubscribe } from "@/webSocketClient";
+import { subscribe } from "@/eventClient";
 import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 
@@ -25,7 +24,7 @@ export default function SystemStatusClient() {
   const { t } = useTranslation();
 
   useEffect(() => {
-    const off = wsSubscribe("jobUpdate", (data) => {
+    const off = subscribe("jobUpdate", (data) => {
       const info = data as JobResponse;
       const filtered =
         filter !== "" ? info.jobs.filter((j) => j.type === filter) : info.jobs;
@@ -33,19 +32,7 @@ export default function SystemStatusClient() {
       setAuditedAt(info.auditedAt);
       setUpdatedAt(info.updatedAt);
     });
-    if (off) return off;
-    const url =
-      filter !== ""
-        ? `/api/system/jobs/stream?type=${encodeURIComponent(filter)}`
-        : "/api/system/jobs/stream";
-    const es = apiEventSource(url);
-    es.onmessage = (e) => {
-      const info: JobResponse = JSON.parse(e.data);
-      setJobs(info.jobs);
-      setAuditedAt(info.auditedAt);
-      setUpdatedAt(info.updatedAt);
-    };
-    return () => es.close();
+    return off;
   }, [filter]);
 
   const types = Array.from(new Set(jobs.map((j) => j.type)));
