+++ src/eventClient.ts
@@ -1,0 +1,10 @@
+import { subscribe as wsSubscribe } from "./webSocketClient";
+
+export type Unsubscribe = () => void;
+
+export function subscribe(
+  event: string,
+  cb: (data: unknown) => void,
+): Unsubscribe | undefined {
+  return wsSubscribe(event, cb);
+}
